fixed bug in targ sweep Navg was float instead of int
Changed Navg import from gc to be np.int instead on np.float
This was causing saveSweepData to crash

+++ kidPy.py
@@ -405,7 +405,7 @@ def targetSweep(ri, udp, valon):
            float span: Sweep span, Hz
            Navg = Number of data points to average at each sweep step"""
     span = np.float(gc[np.where(gc == 'targ_span')[0][0]][1])
-    Navg = np.float(gc[np.where(gc == 'Navg')[0][0]][1])
+    Navg = np.int(gc[np.where(gc == 'Navg')[0][0]][1])
     vna_savepath = str(np.load("last_vna_dir.npy"))
     if not os.path.exists(targ_savepath):
         os.makedirs(targ_savepath)
